+++ shared/src/commonTest/kotlin/com/mbta/tid/mbta_app/model/AlertTest.kt
@@ -70,7 +70,7 @@ class AlertTest {
                 )
             }
         val filteredList =
-            Alert.filter(
+            Alert.applicableAlerts(
                 listOf(
                     alertMatch,
                     alertMatchNoDirection,
@@ -84,6 +84,166 @@ class AlertTest {
             )
 
         assertEquals(listOf(alertMatch, alertMatchNoDirection), filteredList)
+    }
+
+
+    @Test
+    fun `filters applicable alerts`() {
+        val objects = ObjectCollectionBuilder()
+        val stop = objects.stop()
+        val route = objects.route { sortOrder = 1 }
+
+        val validAlert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(
+                        Alert.InformedEntity.Activity.Board,
+                        Alert.InformedEntity.Activity.Exit,
+                        Alert.InformedEntity.Activity.Ride
+                    ),
+                    route = route.id,
+                    routeType = route.type,
+                    stop = stop.id
+                )
+            }
+        val invalidAlert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
+                    route = "wrong",
+                    routeType = route.type,
+                    stop = "wrong"
+                )
+            }
+        assertEquals(
+            Alert.applicableAlerts(      listOf(validAlert, invalidAlert),
+                null,
+                listOf(route.id),
+                setOf(stop.id),
+            ),
+            listOf(validAlert)
+        )
+    }
+
+    @Test
+    fun `filters out alerts without Board activity`() {
+        val objects = ObjectCollectionBuilder()
+        val stop = objects.stop()
+        val route = objects.route { sortOrder = 1 }
+
+        val alert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
+                    route = route.id,
+                    routeType = route.type,
+                    stop = stop.id
+                )
+            }
+        assertEquals(
+            Alert.applicableAlerts(
+                listOf(alert),
+                null,
+                        listOf(route.id),
+                setOf(stop.id),
+            ),
+            emptyList()
+        )
+    }
+
+    @Test
+    fun `filters out alerts with non-matching route ID`() {
+        val objects = ObjectCollectionBuilder()
+        val stop = objects.stop()
+        val route = objects.route { sortOrder = 1 }
+
+        val alert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
+                    route = "not matching",
+                    routeType = route.type,
+                    stop = stop.id
+                )
+            }
+        assertEquals(
+            Alert.applicableAlerts(
+                listOf(alert),
+                null,
+                listOf(route.id),
+                setOf(stop.id),
+            ),
+            emptyList()
+        )
+    }
+
+    @Test
+    fun `filters out alerts with non-matching stop ID`() {
+        val objects = ObjectCollectionBuilder()
+        val stop = objects.stop()
+        val route = objects.route { sortOrder = 1 }
+
+        val alert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
+                    route = route.id,
+                    routeType = route.type,
+                    stop = "not matching"
+                )
+            }
+        assertEquals(
+            Alert.applicableAlerts(
+                listOf(alert),
+                null,
+                listOf(route.id),
+                setOf(stop.id),
+            ),
+            emptyList()
+        )
+    }
+
+    @Test
+    fun `filters to route only when not given set of stops`() {
+        val objects = ObjectCollectionBuilder()
+        val route = objects.route { sortOrder = 1 }
+        val otherRoute = objects.route { sortOrder = 1 }
+
+        val alert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(Alert.InformedEntity.Activity.Board),
+                    route = route.id,
+                    routeType = route.type,
+                    stop = "not matching"
+                )
+            }
+
+        val otherAlert =
+            objects.alert {
+                effect = Alert.Effect.Suspension
+                informedEntity(
+                    listOf(Alert.InformedEntity.Activity.Board),
+                    route = otherRoute.id,
+                    routeType = otherRoute.type,
+                    stop = "not matching"
+                )
+            }
+        assertEquals(
+            listOf(alert),
+            Alert.applicableAlerts(   listOf(alert, otherAlert),
+                null,
+                listOf(route.id),
+                null
+
+            ),
+        )
     }
 
     @Test
