+++ shared/src/commonTest/kotlin/com/mbta/tid/mbta_app/model/RealtimePatternsTest.kt
@@ -593,166 +593,6 @@ class RealtimePatternsTest {
     }
 
     @Test
-    fun `filters applicable alerts`() {
-        val objects = ObjectCollectionBuilder()
-        val stop = objects.stop()
-        val route = objects.route { sortOrder = 1 }
-
-        val validAlert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(
-                        Alert.InformedEntity.Activity.Board,
-                        Alert.InformedEntity.Activity.Exit,
-                        Alert.InformedEntity.Activity.Ride
-                    ),
-                    route = route.id,
-                    routeType = route.type,
-                    stop = stop.id
-                )
-            }
-        val invalidAlert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
-                    route = "wrong",
-                    routeType = route.type,
-                    stop = "wrong"
-                )
-            }
-        assertEquals(
-            RealtimePatterns.applicableAlerts(
-                listOf(route.id),
-                setOf(stop.id),
-                null,
-                listOf(validAlert, invalidAlert)
-            ),
-            listOf(validAlert)
-        )
-    }
-
-    @Test
-    fun `filters out alerts without Board activity`() {
-        val objects = ObjectCollectionBuilder()
-        val stop = objects.stop()
-        val route = objects.route { sortOrder = 1 }
-
-        val alert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
-                    route = route.id,
-                    routeType = route.type,
-                    stop = stop.id
-                )
-            }
-        assertEquals(
-            RealtimePatterns.applicableAlerts(
-                listOf(route.id),
-                setOf(stop.id),
-                null,
-                listOf(alert)
-            ),
-            emptyList()
-        )
-    }
-
-    @Test
-    fun `filters out alerts with non-matching route ID`() {
-        val objects = ObjectCollectionBuilder()
-        val stop = objects.stop()
-        val route = objects.route { sortOrder = 1 }
-
-        val alert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
-                    route = "not matching",
-                    routeType = route.type,
-                    stop = stop.id
-                )
-            }
-        assertEquals(
-            RealtimePatterns.applicableAlerts(
-                listOf(route.id),
-                setOf(stop.id),
-                null,
-                listOf(alert)
-            ),
-            emptyList()
-        )
-    }
-
-    @Test
-    fun `filters out alerts with non-matching stop ID`() {
-        val objects = ObjectCollectionBuilder()
-        val stop = objects.stop()
-        val route = objects.route { sortOrder = 1 }
-
-        val alert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(Alert.InformedEntity.Activity.Exit, Alert.InformedEntity.Activity.Ride),
-                    route = route.id,
-                    routeType = route.type,
-                    stop = "not matching"
-                )
-            }
-        assertEquals(
-            RealtimePatterns.applicableAlerts(
-                listOf(route.id),
-                setOf(stop.id),
-                null,
-                listOf(alert)
-            ),
-            emptyList()
-        )
-    }
-
-    @Test
-    fun `filters to route only when not given set of stops`() {
-        val objects = ObjectCollectionBuilder()
-        val route = objects.route { sortOrder = 1 }
-        val otherRoute = objects.route { sortOrder = 1 }
-
-        val alert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(Alert.InformedEntity.Activity.Board),
-                    route = route.id,
-                    routeType = route.type,
-                    stop = "not matching"
-                )
-            }
-
-        val otherAlert =
-            objects.alert {
-                effect = Alert.Effect.Suspension
-                informedEntity(
-                    listOf(Alert.InformedEntity.Activity.Board),
-                    route = otherRoute.id,
-                    routeType = otherRoute.type,
-                    stop = "not matching"
-                )
-            }
-        assertEquals(
-            listOf(alert),
-            RealtimePatterns.applicableAlerts(
-                listOf(route.id),
-                null,
-                null,
-                listOf(alert, otherAlert)
-            ),
-        )
-    }
-
-    @Test
     fun `properly applies platform alerts by pattern`() {
         val objects = ObjectCollectionBuilder()
         lateinit var platform1: Stop
